ajout de dossier opérationnel

+++ src/main/java/modele/Dossier.java
@@ -15,14 +15,14 @@ public class Dossier {
     private int id_dossier;
 
     private Date dateDossier;
-    private String descriprion;
+    private String description;
     private String nivGravite;
     private int ref_fichepatient;
 
 
 
     public Dossier(String text, String text1, int ref_fichepatient) {
-        this.descriprion=text;
+        this.description=text;
         this.nivGravite=text1;
         this.ref_fichepatient= ref_fichepatient;
     }
@@ -36,34 +36,40 @@ public class Dossier {
     public Dossier(int id_dossier, java.sql.Date dateDossier, String description, String nivGravite, int ref_fichepatient) {
         this.id_dossier = id_dossier;
         this.dateDossier = dateDossier;
-        this.descriprion = description;
+        this.description = description;
         this.nivGravite = nivGravite;
         this.ref_fichepatient = ref_fichepatient;
     }
 
+    public Dossier(String text, String text1) {
+        this.description = text;
+        this.nivGravite =text1;
+    }
 
-    public void ajoutDossierPatient()  throws SQLException {
+
+    public void ajoutDossierPatient(int idFichePatient)  throws SQLException {
         BDD mabdd = new BDD();
         PreparedStatement maRequete = mabdd.getBDD().prepareStatement("INSERT INTO dossier (description,nivGravite,ref_fichepatient) VALUES (?,?,?)");
 
-        maRequete.setString(1, descriprion);
+        maRequete.setString(1, description);
         maRequete.setString(2, nivGravite);
         maRequete.setInt(3, ref_fichepatient);
         int mesResultats = maRequete.executeUpdate();
     }
 
-    public void deleteDossierPatient() throws SQLException {
-        BDD mabdd = new BDD();
-        PreparedStatement maRequete = mabdd.getBDD().prepareStatement("DELETE FROM dossier where id_dossier =?");
-        maRequete.setInt(1, id_dossier );
-        maRequete.executeUpdate();
-
+    public void deleteDossierPatient(Dossier doss) throws SQLException {
+        if(doss.getId_dossier() > 0 ) {
+            BDD mabdd = new BDD();
+            PreparedStatement maRequete = mabdd.getBDD().prepareStatement("DELETE FROM dossier where id_dossier =?");
+            maRequete.setInt(1, id_dossier);
+            maRequete.executeUpdate();
+        }
     }
 
     public void updateDossierPatient() throws SQLException{
         BDD mabdd = new BDD();
-        PreparedStatement maRequete = mabdd.getBDD().prepareStatement("UPDATE dossier SET `descriprion`=?,`nivGravite`=?,`ref_fichepatient`=? WHERE id_dossier=?");
-        maRequete.setString(1, descriprion);
+        PreparedStatement maRequete = mabdd.getBDD().prepareStatement("UPDATE dossier SET `description`=?,`nivGravite`=?,`ref_fichepatient`=? WHERE id_dossier=?");
+        maRequete.setString(1, description);
         maRequete.setString(2, nivGravite);
         maRequete.setInt(3, ref_fichepatient);
         maRequete.setInt(4, id_dossier);
@@ -71,9 +77,9 @@ public class Dossier {
     }
 
 
-    public Dossier(int id_dossier, String descriprion, String nivGravite, int ref_fichepatient) {
+    public Dossier(int id_dossier, String description, String nivGravite, int ref_fichepatient) {
         this.id_dossier = id_dossier;
-        this.descriprion = descriprion;
+        this.description = description;
         this.nivGravite = nivGravite;
         this.ref_fichepatient = ref_fichepatient;
     }
@@ -115,12 +121,12 @@ public class Dossier {
     }
 
 
-    public String getDescriprion() {
-        return descriprion;
+    public String getDescription() {
+        return description;
     }
 
-    public void setDescriprion(String descriprion) {
-        this.descriprion = descriprion;
+    public void setDescription(String description) {
+        this.description = description;
     }
 
     public String getNivGravite() {
